Back_Admissionsite/requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.35.0
django-contrib-comments==1.8.0
django-cors-headers==2.2.0
django-countries==5.3
django-crispy-forms==1.7.2
django-dob-widget==1.1.3
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
httplib2==0.11.3
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.7
Pillow==5.1.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22